Kalória type int helyett decimal lett

+++ NaturaCo_Boundle_api/Models/RecipeIngredientDto.cs
@@ -6,7 +6,7 @@ namespace NaturaCo.RecipeSyncApi.Models
         public string ProductName { get; set; }
         public decimal Quantity { get; set; }
         public string Unit { get; set; }
-        public int? Calories { get; set; }
+        public decimal? Calories { get; set; }
         public decimal? Price { get; set; }
         public decimal? PackageQuantity { get; set; }
         public string PackageUnit { get; set; }
